Fixed mistake in CMakeLists.txt commenting out half of the file with multiline comment, moved Response into rest.h,.cpp from response.h,.cpp.
